# Conflicts:
#	src/main/java/pt/insuranced/main/Main.java
#	src/main/java/pt/insuranced/main/cli.java
#	src/main/java/pt/insuranced/persistence/connection/ConnectionManager.java
#	src/main/resources/LabelsBundle.properties
#	src/main/resources/LabelsBundle_fr_FR.properties

+++ src/main/java/pt/insuranced/models/Password.java
@@ -1,14 +1,14 @@
 package pt.insuranced.models;
 
 public class Password {
-    private int id;
+    private Long id;
 
     private String hashedPassword;
 
     public Password() {
     }
 
-    public Password(int id, String hashedPassword) {
+    public Password(Long id, String hashedPassword) {
         this.id = id;
         this.hashedPassword = hashedPassword;
     }
@@ -21,7 +21,7 @@ public class Password {
         this.hashedPassword = hashedPassword;
     }
 
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
